create api module and refactor forum module
