+++ api/src/main/java/org/discu2/forum/api/packet/ErrorMessagePacket.java
@@ -1,4 +1,4 @@
-package org.discu2.forum.packet;
+package org.discu2.forum.api.packet;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
