+++ api/src/main/java/org/discu2/forum/api/packet/TextBlockRequestPacket.java
@@ -1,4 +1,4 @@
-package org.discu2.forum.packet;
+package org.discu2.forum.api.packet;
 
 import lombok.Data;
 import lombok.Getter;
