+++ forum/build.gradle
@@ -1,7 +1,3 @@
-
-
-
-
 configurations {
     compileOnly {
         extendsFrom annotationProcessor
@@ -22,6 +18,7 @@ dependencies {
 
     implementation 'com.auth0:java-jwt:3.18.1'
     implementation 'junit:junit:4.13.1'
+    implementation project(':api')
 
     testImplementation 'org.springframework.security:spring-security-test'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
