+++ forum/src/main/java/org/discu2/forum/config/MongoGridFsTemplateConfig.java
@@ -1,15 +1,15 @@
-package org.discu2.forum.util;
+package org.discu2.forum.config;
 
 import lombok.AllArgsConstructor;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.MongoDatabaseFactory;
 import org.springframework.data.mongodb.core.convert.MappingMongoConverter;
 import org.springframework.data.mongodb.gridfs.GridFsTemplate;
-import org.springframework.stereotype.Component;
 
-@Component
+@Configuration
 @AllArgsConstructor
-public class MongoGridFsTemplateProvider {
+public class MongoGridFsTemplateConfig {
 
     private MappingMongoConverter mongoConverter;
     private MongoDatabaseFactory mongoDatabaseFactory;
