+++ forum/src/main/java/org/discu2/forum/controller/AccountController.java
@@ -5,12 +5,12 @@ import com.auth0.jwt.exceptions.JWTDecodeException;
 import com.google.common.base.Strings;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.discu2.forum.exception.BadPacketFormatException;
+import org.discu2.forum.api.exception.BadPacketFormatException;
 import org.discu2.forum.model.Account;
-import org.discu2.forum.packet.AccountPacket;
-import org.discu2.forum.packet.AccountUpdateRequestPacket;
-import org.discu2.forum.packet.RegisterRequestPacket;
-import org.discu2.forum.packet.TokenPacket;
+import org.discu2.forum.api.packet.AccountPacket;
+import org.discu2.forum.api.packet.AccountUpdateRequestPacket;
+import org.discu2.forum.api.packet.RegisterRequestPacket;
+import org.discu2.forum.api.packet.TokenPacket;
 import org.discu2.forum.service.AccountService;
 import org.discu2.forum.util.JsonConverter;
 import org.discu2.forum.util.TokenFactory;
