+++ forum/src/main/java/org/discu2/forum/controller/CommentController.java
@@ -1,8 +1,8 @@
 package org.discu2.forum.controller;
 
 import lombok.AllArgsConstructor;
-import org.discu2.forum.exception.BadPacketFormatException;
-import org.discu2.forum.packet.TextBlockRequestPacket;
+import org.discu2.forum.api.exception.BadPacketFormatException;
+import org.discu2.forum.api.packet.TextBlockRequestPacket;
 import org.discu2.forum.service.CommentService;
 import org.discu2.forum.util.JsonConverter;
 import org.springframework.security.access.prepost.PreAuthorize;
