+++ forum/src/main/java/org/discu2/forum/repository/RoleRepository.java
@@ -1,9 +1,8 @@
 package org.discu2.forum.repository;
 
-import org.discu2.forum.model.Role;
+import org.discu2.forum.api.model.Role;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import java.util.Collections;
 import java.util.Optional;
 
 public interface RoleRepository extends MongoRepository<Role, String> {
