+++ forum/src/main/java/org/discu2/forum/repository/TopicRepository.java
@@ -1,12 +1,8 @@
 package org.discu2.forum.repository;
 
-import org.discu2.forum.model.Topic;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.mongodb.core.query.Criteria;
+import org.discu2.forum.api.model.Topic;
 import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.mongodb.repository.Query;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface TopicRepository extends MongoRepository<Topic, String> {
