+++ forum/src/main/java/org/discu2/forum/service/AccountService.java
@@ -5,11 +5,11 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import lombok.AllArgsConstructor;
 import lombok.NonNull;
-import org.discu2.forum.exception.AlreadyExistException;
-import org.discu2.forum.exception.BadPacketFormatException;
-import org.discu2.forum.exception.DataNotFoundException;
+import org.discu2.forum.api.exception.AlreadyExistException;
+import org.discu2.forum.api.exception.BadPacketFormatException;
+import org.discu2.forum.api.exception.DataNotFoundException;
 import org.discu2.forum.model.Account;
-import org.discu2.forum.packet.AccountUpdateRequestPacket;
+import org.discu2.forum.api.packet.AccountUpdateRequestPacket;
 import org.discu2.forum.repository.AccountRepository;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
