+++ forum/src/main/java/org/discu2/forum/service/CommentService.java
@@ -3,11 +3,10 @@ package org.discu2.forum.service;
 import com.google.common.collect.Lists;
 import lombok.AllArgsConstructor;
 import lombok.NonNull;
-import org.discu2.forum.exception.DataNotFoundException;
+import org.discu2.forum.api.model.TextBlock;
+import org.discu2.forum.api.exception.DataNotFoundException;
 import org.discu2.forum.model.Account;
-import org.discu2.forum.model.TextBlock;
 import org.discu2.forum.repository.CommentRepository;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
