+++ forum/src/main/java/org/discu2/forum/service/FileService.java
@@ -4,8 +4,8 @@ import com.google.common.base.Strings;
 import com.mongodb.client.gridfs.model.GridFSFile;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import org.discu2.forum.exception.BadPacketFormatException;
-import org.discu2.forum.exception.DataNotFoundException;
+import org.discu2.forum.api.exception.BadPacketFormatException;
+import org.discu2.forum.api.exception.DataNotFoundException;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.mongodb.gridfs.GridFsTemplate;
 import org.springframework.stereotype.Service;
