+++ forum/src/main/java/org/discu2/forum/service/PostService.java
@@ -3,12 +3,11 @@ package org.discu2.forum.service;
 import com.google.common.collect.Lists;
 import lombok.AllArgsConstructor;
 import lombok.NonNull;
-import org.discu2.forum.exception.DataNotFoundException;
+import org.discu2.forum.api.model.TextBlock;
+import org.discu2.forum.api.exception.DataNotFoundException;
 import org.discu2.forum.model.Account;
-import org.discu2.forum.model.TextBlock;
 import org.discu2.forum.repository.PostRepository;
 import org.springframework.context.annotation.Lazy;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
